feat(dotnet): close every remaining .NET round-trip gap in the fidelity manifest
Fixes the object-callback crash — responseCallback was typed as a
string while the server emits a boolean, so reading such an expectation
back threw and lost the whole object — and corrects refillPerSecond to
decimal so fractional rates keep their scale. Path parameter values
widen to JSON elements so schema-matcher forms deserialize, requests
gain the DNS matcher fields, errors and forwards gain delays, the
forwarded-request override gains the response form, gRPC messages gain
template types, WebSocket responses gain frame matchers, and the LLM
model gains completion extras, token accounting, subword streaming,
rerank, moderation, content filter, and the primary flag. With both
former deserialization throwers fixed, the fidelity test drops its
strip-and-retry scaffolding for a straight round-trip, and the
manifest's dotnet entry is now empty.

+++ mockserver-client-dotnet/src/MockServer.Client/Models/RateLimit.cs
@@ -33,9 +33,14 @@ public sealed class RateLimit
     [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
     public int? Burst { get; set; }
 
+    /// <summary>
+    /// <c>token_bucket</c>: token refill rate per second. Typed as <see cref="decimal"/> (not
+    /// <see cref="double"/>) so a whole-number rate written on the wire as <c>5.0</c> preserves its
+    /// trailing zero on round-trip (a <c>double</c> would re-serialise it as <c>5</c>).
+    /// </summary>
     [JsonPropertyName("refillPerSecond")]
     [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
-    public double? RefillPerSecond { get; set; }
+    public decimal? RefillPerSecond { get; set; }
 
     [JsonPropertyName("errorStatus")]
     [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
